move the autoencoder pytorch model to a separate ae_module file
